fastbootd: Allow using kernel commits to pass SafetyNet
When using kernel commits to pass SafetyNet and CTS like:

  proc: Remove verifiedbootstate flag from /proc/cmdline
  proc: Set androidboot.verifiedbootstate=green

fastbootd returns the device as locked when flashing images
or resizing partitions when flashing because it can't find
androidboot.verifiedbootstate=orange in the kernel command
line letting it know it's safe to flash.

So if androidboot.verifiedbootstate is missing from the kernel
command line or if is returning verifiedbootstate=green report
the device as unlocked.

PS2: Add support for androidboot.verifiedbootstate=green

Change-Id: I56712f374ea26397ea5a183cfe100cf92e21d531
Signed-off-by: Jason Edson <jaysonedson@gmail.com>

Former-commit-id: e4114606c09fd270fa5a57c1c44695f7d4b0a940
Signed-off-by: Jason Edson <jaysonedson@gmail.com>

Former-commit-id: 141f9e4f3f15ff491d1ef5e6e5c27df5565a8156